+++ src/egeo-demo/st-sidebar-demo/st-sidebar-demo.component.ts
@@ -1,0 +1,28 @@
+/*
+ * © 2017 Stratio Big Data Inc., Sucursal en España.
+ *
+ * This software is licensed under the Apache License, Version 2.0.
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the terms of the License for more details.
+ *
+ * SPDX-License-Identifier: Apache-2.0.
+ */
+import { Component } from '@angular/core';
+import { StSidebarItem } from '@stratio/egeo';
+
+@Component({
+   selector: 'st-sidebar-demo',
+   templateUrl: './st-sidebar-demo.component.html',
+   styleUrls: ['./st-sidebar-demo.component.scss']
+})
+
+export class StSidebarDemoComponent {
+   public items: StSidebarItem[] = [
+      {id: 'vault-roles', label: 'Vault Roles'},
+      {id: 'identities', label: 'Identities'},
+      {id: 'masters', label: 'Masters'},
+      {id: 'agents', label: 'Agents'},
+      {id: 'roles', label: 'Roles'}
+   ];
+}