+++ src/lib/barrels.ts
@@ -34,6 +34,7 @@ import { StRadioMenuModule } from './st-radio-menu/st-radio-menu.module';
 import { StRadioModule } from './st-radio/st-radio.module';
 import { StSearchModule } from './st-search/st-search.module';
 import { StSelectModule } from './st-select/st-select.module';
+import { StSidebarModule } from './st-sidebar/st-sidebar.module';
 import { StSpinnerModule } from './st-spinner/st-spinner.module';
 import { StSwitchModule } from './st-switch/st-switch.module';
 import { StTabBoxModule } from './st-tab-box/st-tab-box.module';
@@ -74,6 +75,7 @@ export const DECLARATIONS: any[] = [
    StRadioModule,
    StSearchModule,
    StSelectModule,
+   StSidebarModule,
    StSpinnerModule,
    StSwitchModule,
    StTabBoxModule,